Videoplayer Crash if no preferred tuner set
I came across this issue, when you havent set a preferred tuner in settings, the videoplayer will crash in the ss.good() loop. Ive just added check if preferred tuner is empty (none set) the function will continue on. Im not sure on the best way to continue, so feel free to provide any alternative solution.
With this change, the Videoplayer will start and run the channel as intended